Move cni commands to `pkg/cnicommands` package 2/2
